+++ src/main/java/com/retos/rentacar/modelo/DTO/ClientAndKeyclient.java
@@ -1,7 +1,7 @@
-package com.retos.rentacar.modelo.DTO;
+package com.retos.rentacar.servicios.modelo.DTO;
 
-import com.retos.rentacar.modelo.Entity.Client.Client;
-import com.retos.rentacar.modelo.Entity.Client.KeyClient;
+import com.retos.rentacar.servicios.modelo.Entity.Client.Client;
+import com.retos.rentacar.servicios.modelo.Entity.Client.KeyClient;
 
 public class ClientAndKeyclient {
