+++ src/main/java/com/retos/rentacar/modelo/Entity/Message/Message.java
@@ -1,9 +1,10 @@
 
-package com.retos.rentacar.modelo.Entity.Message;
+package com.retos.rentacar.servicios.modelo.Entity.Message;
 
+import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.retos.rentacar.modelo.Entity.Car.Car;
-import com.retos.rentacar.modelo.Entity.Client.Client;
+import com.retos.rentacar.servicios.modelo.Entity.Car.Car;
+import com.retos.rentacar.servicios.modelo.Entity.Client.Client;
 
 import java.io.Serializable;
 import java.util.Optional;
@@ -20,13 +21,16 @@ public class Message implements Serializable {
     @Column(name = "message_text")
     private String messageText;
 
+
+    @JsonIgnore
     @ManyToOne
-    @JoinColumn(name = "id_car")
+    @JoinColumn(name = "car_id")
     @JsonIgnoreProperties({ "messages", "client", "reservations" })
     private Car car;
 
+    @JsonIgnore
     @ManyToOne
-    @JoinColumn(name = "id_client")
+    @JoinColumn(name = "client_id")
     @JsonIgnoreProperties({ "messages", "reservations", "client" })
     private Client client;
