+++ src/main/java/com/retos/rentacar/modelo/Entity/Reservation/ReservationCode.java
@@ -1,4 +1,4 @@
-package com.retos.rentacar.modelo.Entity.Reservation;
+package com.retos.rentacar.servicios.modelo.Entity.Reservation;
 
 import java.util.Random;
